chore: migrate xor and utils
